Move ObjectBox struct wrappers to separate file
